requirements.txt
asgiref==3.5.0
astroid==2.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.2
django-allauth==0.47.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
oauthlib==3.2.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
wrapt==1.13.3
